refactor: 함수가 하나의 기능만 수행하도록 변경하라
Application의 main 메서드가 우승자 명단을 추출하는 로직을 서비스 계층에 위임하였습니다.
그 결과, main 메서드는 프로그램 실행에 필요한 객체를 생성하고 메서드를 호출하는 기능만 수행합니다.
