+++ src/main/java/racingcar/car/InputGenerator.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.car;
 
 public interface InputGenerator {
