+++ src/test/java/racingcar/car/CarNameTest.java
@@ -1,8 +1,9 @@
-package racingcar;
+package racingcar.car;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import racingcar.car.CarName;
 
 class CarNameTest {
