+++ src/test/java/racingcar/car/CarsTest.java
@@ -1,11 +1,12 @@
-package racingcar;
+package racingcar.car;
 
-import camp.nextstep.edu.missionutils.Randoms;
 import java.util.List;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import racingcar.car.Car;
+import racingcar.car.Cars;
 
 class CarsTest {
